Moved Hub screen components into separate files.
