+++ iOS/MobileStarter/react-app/src/Exports.tsx
@@ -15,6 +15,10 @@ export * from "./InfoBottomPanel";
 export * from "./AboutBottomPanel";
 export * from "./ViewsBottomPanel";
 export * from "./HomeScreen";
+export * from "./HubScreenButton";
+export * from "./IModelDownloader";
+export * from "./IModelPicker";
+export * from "./ProjectPicker";
 export * from "./HubScreen";
 export * from "./ElementPropertiesPanel";
 export * from "./ToolsBottomPanel";
